client/requirements.txt
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
docutils==0.18.1
ffpyplayer==4.3.5
Flask==2.2.1
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Kivy==2.1.0
Kivy-Garden==0.1.5
MarkupSafe==2.1.1
Pillow==9.1.0
pycparser==2.21
Pygments==2.12.0
PyNaCl==1.5.0
requests==2.27.1
urllib3==1.26.9
Werkzeug==2.2.1
zipp==3.8.1
